[lib/MSSim] performance optimization
Calculate all the trig ratio functions up-front.
This make us rely less on the compiler CSE which doesn't seem to work
too well for code with branches.
(OTOH, DCE when we are not using a particular value should be quite good)

Passing them into the kernel with a macro-generated structure
so that we can still share as much beam path as possible between
the test version and the actually used version.